+++ src/com/syncleus/tests/dann/genetics/TestByteValueGene.java
@@ -27,13 +27,13 @@ public class TestByteValueGene
 	public void testConstructors()
 	{
 		ValueGene test = new ByteValueGene((byte) 123);
-		Assert.assertTrue("value constructor failed", test.getValue().getNumber().byteValue() == 123);
+		Assert.assertTrue("value constructor failed", test.getValue().getNumber().byteValue() == (byte) 123);
 		test = new ByteValueGene(new MutableByte((byte) 57));
-		Assert.assertTrue("MutableByte value constructor failed", test.getValue().getNumber().byteValue() == 57);
-		test = new ByteValueGene(Byte.valueOf((byte) 83));
-		Assert.assertTrue("Number value constructor failed", test.getValue().getNumber().byteValue() == 83);
+		Assert.assertTrue("MutableByte value constructor failed", test.getValue().getNumber().byteValue() == (byte) 57);
+		test = new ByteValueGene((byte) 83);
+		Assert.assertTrue("Number value constructor failed", test.getValue().getNumber().byteValue() == (byte) 83);
 		test = new ByteValueGene();
-		Assert.assertTrue("default constructor failed", test.getValue().getNumber().byteValue() == 0);
+		Assert.assertTrue("default constructor failed", test.getValue().getNumber().byteValue() == (byte) 0);
 	}
 
 	@Test
