+++ src/com/syncleus/tests/dann/genetics/wavelets/TestSignalProcessingWavelet.java
@@ -102,7 +102,7 @@ public class TestSignalProcessingWavelet
 			while (population.size() < POPULATION_SIZE)
 			{
 				SignalProcessingWavelet processor;
-				if (RANDOM.nextFloat() < 0.5)
+				if (RANDOM.nextFloat() < (float) 0.5)
 					processor = populationArray.get(RANDOM.nextInt(populationArray.size()));
 				else
 					processor = new SignalProcessingWavelet(xAxis, output);
